Linters and black

+++ src/pynaneye/frame.py
@@ -8,12 +8,13 @@ from numpy._typing import NDArray
 def frame_bytes_to_array(frame_bytes: bytes, w: int, h: int) -> NDArray[uint8]:
     """Converts raw frame bytes to a NumPy array."""
     bayer_data = frombuffer(bytearray(frame_bytes), dtype=uint8).reshape((h, w, 3))
-    return cv2.cvtColor(bayer_data, cv2.COLOR_RGB2BGR)
+    return cv2.cvtColor(bayer_data, cv2.COLOR_RGB2BGR).astype(uint8)
 
 
 @dataclass
 class NanEyeFrame:
     """Dataclass to hold NanEye frame data."""
+
     image_bytes: bytes
     width: int
     height: int
